requirements.txt
google-api-python-client
pandas
pusher
squareup
